refactor: migrate core home page to a template
